+++ guillotina/db/storages/pg.py
@@ -482,6 +482,16 @@ class PostgresqlStorage(BaseStorage):
             raise KeyError(oid)
         return objects
 
+    def log_conflict_error(self, oid, txn, old_serial, writer, message='Conflict'):
+        log.warn(
+            f'{message}\n'
+            f'Object ID: {oid}\n'
+            f'TID: {txn._tid}\n'
+            f'Old Object TID: {old_serial}\n'
+            f'Belongs to: {writer.of}\n'
+            f'Parent ID: {writer.id}'
+        )
+
     async def store(self, oid, old_serial, writer, obj, txn):
         assert oid is not None
 
@@ -520,20 +530,28 @@ class PostgresqlStorage(BaseStorage):
                 )
             except asyncpg.exceptions.ForeignKeyViolationError:
                 txn.deleted[obj._p_oid] = obj
+                self.log_conflict_error(oid, txn, old_serial, writer,
+                                        'TID Conflict bad value')
                 raise TIDConflictError(
                     'Bad value inserting into database that could be caused '
                     'by a bad cache value. This should resolve on request retry.')
             except asyncpg.exceptions._base.InterfaceError as ex:
                 if 'another operation is in progress' in ex.args[0]:
+                    self.log_conflict_error(oid, txn, old_serial, writer,
+                                            'Conflict another op in progress')
                     raise ConflictError(
                         'asyncpg error, another operation in progress.')
                 raise
             except asyncpg.exceptions.DeadlockDetectedError:
+                self.log_conflict_error(oid, txn, old_serial, writer,
+                                        'Conflict: Deadlock')
                 raise ConflictError(
                     'Deadlock detected.')
             if len(result) != 1 or result[0]['count'] != 1:
                 if update:
                     # raise tid conflict error
+                    self.log_conflict_error(oid, txn, old_serial, writer,
+                                            'TID Mismatch')
                     raise TIDConflictError(
                         'Mismatch of tid of object being updated. This is likely '
                         'caused by a cache invalidation race condition and should '